Reduce race condition for unlinks.
If multiple clients are active at once, the following could occur:

* Client #1 unlinks file A
* Client #2 creates file B, which reuses A's inode
* Client #2 sends request to server
* Client #1 sends request to server
* Processing client #2's request creates a mismatch warning for
  file A/B.
* Processing client #1's request creates a mismatch warning too.

Note that this can happen even if Client #2 sends its request later,
as there's no intrinsic guarantee of the order in which requests
are processed; any SINGLE client is presumably executing operations
in order, but multiple clients aren't.

Fixing this in rmdir, unlink, and rename.