Don't report inode or path mismatches on OP_EXEC.
We don't canonicalize OP_EXEC names (because they don't
want to be adapted for the chroot environment -- maybe we
should be doing that anyway, but right now we're not), so
mismatches with them are meaningless.  So are mismatches
with the inode 0 reported for something that we never tried
to stat.