Updates/improvements for logging
Send program name (program_invocation_name from glibc) along with the
tag.

Along the way, restructure the fds/pids/tags arrays to be an array
of client structures in pseudo_server, and add the message type
to the set of things logged -- logging that a message was a ping is
more useful than appending the text "ping" to it.  Add support
for type and program to pseudolog.

Add deletion to pseudolog.

Handle usage message formatting when there's an odd number of known
specifiers for pseudolog.

Conflicts:

	ChangeLog.txt
	pseudo_server.c