Provide a hook to disable symlink "fixing".
In fakechroot, which pseudo tries to match the functionality of,
the default behavior when creating a symlink with an absolute target
is to prepend the chroot path, so that underlying syscalls will
get the right file.

It is necessary to be able to disable this behavior to create target
filesystems in some cases.  To that end, support a new environment
variable, PSEUDO_NOSYMLINKEXP, which disables that behavior.