Remove PSEUDO_SUFFIX from source
The PSEUDO_SUFFIX thing is an installation quirk to allow our
build system to tag libpseudo.so with a checksum of the host libc.
However, we reuse a prebuilt pseudo server with the new pseudo
libraries; this means that encoding the suffix in the environment
hackery is a Bad Idea.

Update version number to 0.3, since this seems to wrap up a
hunk of development effort.