Handle insane boundary condition for regcomp()/regexec()
For reasons that may never be known, the /usr/bin/find on
RHEL5 contains its own local copies of regcomp() and
regexec().  Thus, when pseudo makes calls to these functions,
it gets the local copies in the binary instead of the ones in
libc.

But wait!  That's not all.  There's also the fascinating
detail that, for reasons unknown, these local copies have
an incompatible API, such that:
	regexec(pattern, list, 1, pmatch, 0);
can write to more than one element of pmatch, and since
that's a local array of one member, that means that they
can crush something on the stack, such that a couple of
function calls later you get a segfault in Nothing In
Particular.

So.  We try to grab the real regcomp/regexec from libc,
using dlsym, and if we can't, we fall back on whatever the
defaults were.

Inexplicably, this code actually made pseudo crash less often
on at least one target.  This is madness within madness, and
I really have no idea why on earth /usr/bin/find, on a Linux
system, would have its own local copies of regcomp/regexec,
let alone why it would have copies that had substantially
different semantics.