Major change:  Replace the shell-based makewrappers with a Python one.
There's a long story here, but to abbreviate it:

The shell script was annoying at best to maintain and starting to
show signs of not really being the right tool for the job.  For
various reasons, we have some other Python stuff in our build system,
so we picked Python as the language we were already using for other
stuff.  We think this works with anything reasonably recent (around
Python 2.4 through 2.6).

There's a little bit of cleanup, also, of the wrapper templates.