Add sqlite call profiling
Add sqlite call profiling, this allows us to see the sqlite calls
that are being made as the system runs, via the pseudo log.

It was noted that by this profiling that a small change to pseudo.c,
when a file was found, reduced the sqlite SELECT calls by about 1/3.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>