Add a cache of the last object found in pseudo_op
Add a cache of the last object found in pseudo_op.  Profiling has indiciated
that many operations come in clusters.  So instead of doing select, operation
for each item in the cluster, we check to see if we already know the item and
perform the op..

Performance improvement when processing 500k or so files:

Previous:
real    7m11.778s
user    0m35.929s
sys     2m46.723s

This commit:
real    6m42.093s
user    0m34.321s
sys     2m46.086s

Also validation of the component can be added by compiling with NVALIDATE.
This verifies the result of the cache is the same as what would have come
from the database.  Differences are logged to the standard pseudo.log.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>