Fixup:  The path code could double-free.
The problem is that path_by_ino could end up being the same pointer
as cache_path, after which, if cache_path were freed (or kept around
for later), there would be malloc arena problems.

Also, fix the calculation for pathlen to increase cache hits.  The
IPC messages use length of path *plus one* as the length, because
the buffer is defined to include its terminating null byte.