Automatically create state/prefix directories
It'd be handy for the WR build system if new state directories could
be created as needed.  It is made so.  And to answer the first
question everyone, including me, has on reading this:  You can't
do system("mkdir -p ...") because the invoked shell would need to
run under pseudo, so it'd have to check for a server, and...