Merge in ports work
This is a spiffied-up rebase of a bunch of intermediate changes, presented
as a whole because it is, surprisingly, less confusing that way.  The basic
idea is to separate the guts code into categories ranging from generic
stuff that can be the same everywhere and specific variants.  The big scary
one is the Darwin support, which actually seems to run okay on 64-bit OS X
10.6.  (No other variants were tested.)  The other example given is support
for the old clone() syscall on RHEL 4, which affects some wrlinux use cases.

There's a few minor cleanup bits here, such as a function with inconsistent
calling conventions, but nothing really exciting.