Fix realpath(name, NULL) when PSEUDO_DISABLED=1
On some Linux systems, dlsym("realpath", RTLD_NEXT) prefers
for reasons of its own to give a symbol that is also known
as old_realpath, which fails and yields EINVAL when called
with a null pointer as the second argument.  This can be
avoided, on some systems, by using dlvsym() to request
the GLIBC_2.3 version of the symbol.

The wrapper logic is enhanced to allow for specifying
versions, although this currently only works for Linux
(Darwin has no dlvsym, apparently?).  The test case is
a trivial program which calls realpath(name, NULL) run
with PSEUDO_DISABLED=1.