Fix uninitalized variable.
You might be wondering why this wasn't caught.  Answer:  gcc's too smart.

Consider:

	int x;

	if (condition)
		x = 23;

	return x;

This function will just return 23.  Since gcc knows that it doesn't
matter what happens if x is used uninitialized (it's an indeterminate
value, thus a possible trap representation, thus undefined behavior
to use it), it simplifies the initial part of this away.  Thus there's
no use of an uninitialized value.

Something similar seems to be at issue with the use of the uninitialized
f in pseudo_init_one_wrapper.

The variable wasn't initialized in the pre-realpath-fix version either,
but in that version, the assignment from dlsym was completely unconditional.