Fix *at() function interface holes
1.  Fix *at() where dirfd is obtained through dirfd(DIR *).

The dirfd(DIR *) interface allows you to get the fd for a DIR *,
meaning you can use it with openat(), meaning you can need its
path.  This causes a segfault.  Also fixed the base_path
code not to segfault in that case, but first fix the
underlying problem.

2.  Implement renameat()

After three long years, someone tried to use this.  This was impossibly
hard back when pseudo was written, because there was only one dirfd
provided for.  Thing is, now, the canonicalization happens in wrapfuncs,
so a small tweak to makewrappers to recognize that oldpath should use
olddirfd if it exists is enough to get us fully canonicalized paths
when needed.