add popen() call
We weren't trapping popen(), so if environment variables were in an
inconsistent state when popen() was called, Bad Things Happened.  Add
a popen() wrapper.  Like a couple of other special cases, is applied
even when pseudo is theoretically disabled, and that includes the antimagic
case.  (But we never use popen() so that's fine.)