Improve RPATH logic
The existing behavior was to set rpath to whatever was specified
explicitly with --with-rpath, or to set a default if the opt_rpath
variable was unset and we reached a --with-sqlite.

This turns out to be incorrect in the case where a static sqlite is
being used.  You can force the issue with --without-rpath, but it's
probably better to make the inference smarter.  This also allows the
slight cleanup of setting opt_rpath to '' to begin with, because we're
no longer depending on the distinction between empty and unset.