Makefile.in/configure: Use $(LIB), not hardcoded lib, for sqlite
It turns out that the -L usage mostly doesn't matter (usually something
else has requested the right directory, or it's the default), but the
explicit path to libsqlite3.a hardcoded "lib", and on some systems
it should be something else, such as "lib64". Solution: Use $(LIB) for
that directory.

Note that this may not resolve things if, say, you're doing MIPS n32 on a
target where that lives in /usr/lib32, but I think in that case you'd
be specifying $libdir, so it should still work out.

Also added --with-static-sqlite=/path as an option in case people need
to further outsmart this.

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>