Fix link() to work when chrooted
The automatic path fixups invoked for names which end in the string
"path" was still applying to link(), which then called linkat(),
which would do the same path fixups; if you were chrooted, this would
produce bogus paths. On systems which actually have linkat(), this
would produce the even more mysterious behavior that the link would
succeed, but the following stat would fail.

Solution: Change the wrapfuncs prototypes for link() so it doesn't
invoke automatic path name fixups.