Extended attribute improvements (no path, binary data)
Issue #1: If an operation came in for an item with no path
provided by the wrapper, the client would not construct the
combined "path" value. Fixed, and missing paths are now
consistently handled as 0-byte paths.

Issue #2: The database code was assuming the values were
strings, and ignoring a specified length.

Issue #3: The computation of the length of the stored value
was off by one, because it was including the extra terminating
null the client added in case the value was a path.

With this in place, "cp -a" on CentOS is consistently
duplicating the system.posix_acl_access fields as expected,
but unfortunately not handling their permissions too.

(Intent is to translate a system.posix_acl_access setxattr
into corresponding permissions whenever possible.)

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>