Extended attributes: Intercept posix_acl_access
In the fairly common case where someone is using
setxattr() to specify the "posix_acl_access" attribute, but in fact
the ACL list specified can be fully represented in a plain old mode,
we intercept the request and just do a chmod. Even if the request
can't be fully represented, we try to represent any aspects of it that
we can in the plain old mode.

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>