xattr support and other path stuff: reduce allocation and copying
The xattr first-pass implementation was allocating a buffer to
hold the name and value for a set operation, then pseudo_client was
allocating *another* buffer to hold the path and those two values.

pseudo_client_op develops more nuanced argument handling, and also
uses a static buffer for the extended paths it sometimes needs. So
for the typical use case, only occasional operations will need to
reallocate/expand the buffer, and we'll be down to copying things
into that buffer once per operation, instead of having two alloc/free
pairs and two copies.

And of course, that wasn't two alloc/free pairs, it was one alloc/free
pair and one alloc without a free. Whoops.

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>