pseudo_has_unload: add function
Various wrappers checked for a non-null pseudo_get_value("PSEUDO_UNLOAD") to
determine whether the environment should include the pseudo variables.  None
of those checks freed the returned value when it was not null.  The new
check function does.

The new check function also sees whether PSEUDO_UNLOAD was defined in the
environment that should be used in the wrapped system call.  This allows
pkg_postinst scripts to strip out the LD_PRELOAD setting, for example before
invoking qemu to execute commands in an environment that does not have
libpseudo.so.

[YOCTO #4843]

Signed-off-by: Peter A. Bigot <pab@pabigot.com>
Signed-off-by: Peter Seebach <peter.seebach@windriver.com>